Update db schemas (#4)
* Add Uniqueness Constraints for DateRange, Entry, Tag

DateRange: start/end dates have to be unique
Entry: written date and subject date have to be unique
Tag: name has to be unique

* Make DateRange.title optional

* Set default for DateRange.events to empty string

See https://github.com/typeorm/typeorm/issues/1532, since simple-array is stored as text under the hood, we need to set the default for `events` to an empty string, rather than an empty array.

* Add cascades

Set most cascades to ['insert', 'update'] to reduce risk of data loss if a single deletion causes a TON of data to be deleted. Exception: deleting a `Tag` deletes all the associated `Note`s.